requirements/base.txt
absl-py==1.0.0
aiobotocore==2.2.0
aiohttp==3.8.1
aiohttp-retry==2.4.6
aioitertools==0.10.0
aiosignal==1.2.0
anyconfig==0.9.11
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
asyncssh==2.10.1
asynctest==0.13.0
atpublic==3.0.1
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blessings==1.7
boto3==1.21.21
botocore==1.24.21
bravado==11.0.3
bravado-core==5.17.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
colorlog==4.2.1
commonmark==0.9.1
configobj==5.0.6
cryptography==36.0.2
cycler==0.11.0
Cython==0.29.24
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
dpath==2.0.6
dulwich==0.20.35
dvc==2.10.1
dvc-render==0.0.4
einops==0.3.2
entrypoints==0.4
faiss-cpu==1.7.1.post2
fastjsonschema==2.15.3
fasttext==0.9.3
filelock==3.6.0
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.32.0
frozenlist==1.3.0
fsspec==2022.3.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
fuzzywuzzy==0.18.0
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.35.0
google-auth-oauthlib==0.4.6
gpustat==0.6.0
grandalf==0.6
grpcio==1.44.0
huggingface-hub==0.5.1
idna==3.3
imageio==2.9.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
ipykernel==6.13.0
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
jmespath==1.0.0
joblib==1.1.0
jsonpointer==2.3
jsonref==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
loguru==0.5.3
mailchecker==4.1.15
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.3.0
matplotlib-inline==0.1.3
mecab-python3==1.0.5
mistune==0.8.4
monotonic==1.6
msgpack==1.0.3
multidict==6.0.2
munch==2.5.0
nanotime==0.5.2
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
neptune-client==0.16.1
nest-asyncio==1.5.5
networkx==2.6.3
nltk==3.7
notebook==6.4.11
numpy==1.19.1
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.0.0.21
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.1.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.46
pickleshare==0.7.5
Pillow==6.2.2
plotly==5.3.1
pluggy==0.13.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pyclipper==1.2.1
pycparser==2.21
pydot==1.4.2
pygit2==1.9.1
Pygments==2.11.2
pygtrie==2.4.2
PyJWT==2.3.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==6.2.5
python-benedict==0.25.0
python-dateutil==2.8.2
python-decouple==3.6
python-fsutil==0.6.0
python-slugify==6.1.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rfc3987==1.3.8
rich==12.2.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3fs==2022.3.0
s3transfer==0.5.2
sacremoses==0.0.49
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
scmrepo==0.0.16
Send2Trash==1.8.0
sentence-transformers==2.1.0
sentencepiece==0.1.96
Shapely==1.7.1
shortuuid==1.0.8
shtab==1.5.4
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
strict-rfc3339==0.7
swagger-spec-validator==2.7.4
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.4.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
terminado==0.13.3
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
tokenizers==0.10.3
toml==0.10.2
torch==1.9.1
torchvision==0.10.1
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
transformers==4.11.3
typing_extensions==4.2.0
unidic-lite==1.0.8
urllib3==1.26.9
voluptuous==0.13.1
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
widgetsnbextension==3.6.0
wrapt==1.14.0
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.8.0
